requirements.txt
flask_script==2.0.6   
flask_migrate== 2.5.3  
flask_sqlalchemy== 2.4.4
pymysql==0.10.1
flask_bootstrap==3.3.7.1
cryptography==3.2.1  
psutil
flask_wtf
pillow
requests
simplejson
pyyaml 
ua-parser 
user-agents